// packages/client/src/components/domain/home/HomePage.module.css.d.ts
declare const styles:
  & Readonly<{ "header": string }>
  & Readonly<{ "navbar": string }>
  & Readonly<{ "logo": string }>
  & Readonly<{ "main": string }>
  & Readonly<{ "intro": string }>
  & Readonly<{ "introTitle": string }>
  & Readonly<{ "introSubtitle": string }>
  & Readonly<{ "introStack": string }>
  & Readonly<{ "introStackButtons": string }>
  & Readonly<{ "reference": string }>
  & Readonly<{ "referenceHeading": string }>
  & Readonly<{ "referenceContents": string }>
;
export default styles;
//# sourceMappingURL=./HomePage.module.css.d.ts.map